requirements.txt
discord.py==2.3.1
python-dotenv==1.0.0
asgiref==3.6.0
openai==0.27.8
revChatGPT==6.8.6
GoogleBard==2.1.0
EdgeGPT==0.13.2
selenium==4.9.1
undetected-chromedriver==3.4.7
requests==2.31.0